requirements.txt
fastapi
uvicorn
jinja2
mysql-connector-python
requests
beautifulsoup4
numpy
matplotlib
scikit-learn
